feat: add 2 pages, js divided into 3 modules, added a day progress indicator
