--- index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-    <link rel="stylesheet" href="style.css" />
-  </head>
-
-  <body>
-    <div class="container">
-      <div class="taskAdding">
-        <input type="text" class="inputAddTask" placeholder="Add new task" />
-        <button class="buttonAddTask">+</button>
-      </div>
-      <div class="taskWidget"></div>
-    </div>
-    <h1 id="dateHeading"></h1>
-    <div class="weekProgressWidget"></div>
-    <script src="script.js"></script>
-  </body>
-</html>
